@@ -1,50 +1,54 @@
 {
   "private": true,
   "name": "tgui-workspace",
-  "version": "4.5.0-paradise",
-  "packageManager": "yarn@4.1.1",
+  "version": "4.5.1-paradise",
+  "packageManager": "yarn@4.3.1",
   "workspaces": [
     "packages/*"
   ],
   "dependencies": {
-    "@babel/core": "^7.23.3",
-    "@babel/eslint-parser": "^7.23.3",
-    "@babel/plugin-transform-class-properties": "^7.23.3",
-    "@babel/plugin-transform-jscript": "^7.23.3",
-    "@babel/preset-env": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
-    "@types/jest": "^29.5.11",
-    "@types/jsdom": "^21.1.6",
-    "@types/node": "^20.10.6",
+    "@babel/core": "^7.24.9",
+    "@babel/eslint-parser": "^7.24.8",
+    "@babel/plugin-transform-class-properties": "^7.24.7",
+    "@babel/plugin-transform-jscript": "^7.24.7",
+    "@babel/preset-env": "^7.24.8",
+    "@babel/preset-typescript": "^7.24.7",
+    "@eslint/compat": "^1.1.1",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.7.0",
+    "@types/jest": "^29.5.12",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^20.14.11",
     "@types/webpack": "^5.28.5",
-    "@types/webpack-env": "^1.18.4",
-    "@typescript-eslint/parser": "^6.18.0",
+    "@types/webpack-env": "^1.18.5",
+    "@typescript-eslint/parser": "^8.0.0-alpha",
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
     "babel-plugin-inferno": "^6.7.1",
     "babel-plugin-transform-remove-console": "^6.9.4",
     "common": "workspace:*",
-    "css-loader": "^6.8.1",
-    "esbuild-loader": "^4.0.2",
-    "eslint": "^8.56.0",
+    "css-loader": "^7.1.2",
+    "esbuild-loader": "^4.2.2",
+    "eslint": "^9.7.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "5.1.2",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-unused-imports": "^3.0.0",
+    "eslint-plugin-prettier": "5.2.1",
+    "eslint-plugin-react": "^7.34.4",
+    "eslint-plugin-unused-imports": "^4.0.0",
     "file-loader": "^6.2.0",
+    "globals": "^15.8.0",
     "jest": "^29.7.0",
     "jest-circus": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "jsdom": "^23.1.0",
-    "mini-css-extract-plugin": "^2.7.6",
-    "prettier": "3.2.5",
-    "sass": "^1.69.7",
-    "sass-loader": "^13.3.2",
-    "style-loader": "^3.3.3",
+    "jsdom": "^24.1.0",
+    "mini-css-extract-plugin": "^2.9.0",
+    "prettier": "3.3.3",
+    "sass": "^1.77.8",
+    "sass-loader": "^14.2.1",
+    "style-loader": "^4.0.0",
     "terser-webpack-plugin": "^5.3.10",
-    "typescript": "^5.3.3",
-    "webpack": "^5.89.0",
-    "webpack-bundle-analyzer": "^4.10.1",
+    "typescript": "^5.5.3",
+    "webpack": "^5.93.0",
+    "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4"
   }
 }
