@@ -33,7 +33,7 @@ export class Countdown extends Component {
   componentDidUpdate(prevProps) {
     if (this.props.current !== prevProps.current) {
       // https://github.com/yannickcr/eslint-plugin-react/issues/1707
-      // eslint-disable-next-line react/no-did-update-set-state
+
       this.setState((prevState) => {
         return {
           value: Math.max(this.props.timeLeft * 100, 0),
