@@ -1,5 +1,5 @@
 import { createPopper, OptionsGeneric } from '@popperjs/core';
-import { Component, findDOMfromVNode, InfernoNode, render } from 'inferno';
+import { Component, findDOMFromVNode, InfernoNode, render } from 'inferno';
 import type * as CSS from 'csstype';
 
 type PopperProps = {
@@ -43,7 +43,7 @@ export class Popper extends Component<PopperProps> {
         // This code is copied from `findDOMNode` in inferno-extras.
         // Because this component is written in TypeScript, we will know
         // immediately if this internal variable is removed.
-        findDOMfromVNode(this.$LI, true),
+        findDOMFromVNode(this.$LI, true),
         this.renderedContent,
         options
       );
