@@ -48,14 +48,14 @@ export const DisposalBin = (props, context) => {
                 icon="toggle-off"
                 disabled={data.isAI || data.panel_open}
                 content="Disengaged"
-                selected={data.flushing ? null : 'selected'}
+                selected={!data.flushing}
                 onClick={() => act('disengageHandle')}
               />
               <Button
                 icon="toggle-on"
                 disabled={data.isAI || data.panel_open}
                 content="Engaged"
-                selected={data.flushing ? 'selected' : null}
+                selected={data.flushing}
                 onClick={() => act('engageHandle')}
               />
             </LabeledList.Item>
@@ -64,14 +64,14 @@ export const DisposalBin = (props, context) => {
                 icon="toggle-off"
                 disabled={data.mode === -1}
                 content="Off"
-                selected={data.mode ? null : 'selected'}
+                selected={!data.mode}
                 onClick={() => act('pumpOff')}
               />
               <Button
                 icon="toggle-on"
                 disabled={data.mode === -1}
                 content="On"
-                selected={data.mode ? 'selected' : null}
+                selected={data.mode}
                 onClick={() => act('pumpOn')}
               />
             </LabeledList.Item>
