@@ -135,7 +135,7 @@ const DroneList = (props, context) => {
         icon="broadcast-tower"
         disabled={ping_cd || !drones.length}
         title={drones.length ? null : 'No active drones!'}
-        fluid={1}
+        fluid
         textAlign="center"
         py={0.4}
         mt={0.6}
