@@ -51,13 +51,14 @@ export const PanDEMIC = (props, context) => {
     <Window width={575} height={510}>
       <Window.Content>
         <Stack fill vertical>
-          {emptyPlaceholder && (
-            <Section title="Container Information" buttons={<CommonCultureActions />}>
+          {emptyPlaceholder && !beakerContainsVirus ? (
+            <Section title="Container Information" buttons={<CommonCultureActions fill vertical />}>
               <NoticeBox>{emptyPlaceholder}</NoticeBox>
-              {resistances?.length > 0 && <ResistancesSection />}
             </Section>
+          ) : (
+            <CultureInformationSection />
           )}
-          {!!beakerContainsVirus && <CultureInformationSection />}
+          {resistances?.length > 0 && <ResistancesSection align="bottom" />}
         </Stack>
       </Window.Content>
     </Window>
@@ -201,11 +202,9 @@ const CultureInformationSection = (props, context) => {
 
   if (strains.length === 0) {
     return (
-      <Stack fill vertical>
-        <Section title="Container Information" buttons={<CommonCultureActions />}>
-          <NoticeBox>No disease detected in provided blood sample.</NoticeBox>
-        </Section>
-      </Stack>
+      <Section title="Container Information" buttons={<CommonCultureActions />}>
+        <NoticeBox>No disease detected in provided blood sample.</NoticeBox>
+      </Section>
     );
   }
 
