@@ -4,6 +4,7 @@ import { resolveAsset } from '../assets';
 import { useBackend } from '../backend';
 import { Box, Button, Icon, Stack } from '../components';
 import { Window } from '../layouts';
+import { InfernoNode } from 'inferno';
 
 const ROWS = 5;
 const COLUMNS = 9;
@@ -23,7 +24,7 @@ const getGridSpotKey = (spot: [number, number]): GridSpotKey => {
   return `${spot[0]}/${spot[1]}`;
 };
 
-const CornerText = (props: { align: 'left' | 'right'; children: string }): JSX.Element => {
+const CornerText = (props: { align: 'left' | 'right'; children: string }): InfernoNode => {
   const { align, children } = props;
 
   return (
@@ -89,7 +90,7 @@ const SLOTS: Record<
     displayName: string;
     gridSpot: GridSpotKey;
     image?: string;
-    additionalComponent?: JSX.Element;
+    additionalComponent?: InfernoNode;
   }
 > = {
   eyes: {
@@ -231,7 +232,7 @@ const ALTERNATIVE_SLOTS: Record<
     displayName: string;
     gridSpot: GridSpotKey;
     image?: string;
-    additionalComponent?: JSX.Element;
+    additionalComponent?: InfernoNode;
   }
 > = {
   eyes: {
