@@ -4,7 +4,6 @@
  * @license MIT
  */
 
-import { Placement } from '@popperjs/core';
 import { Box, Button, Section, Tooltip } from '../components';
 
 export const meta = {
@@ -13,6 +12,7 @@ export const meta = {
 };
 
 const Story = () => {
+  /** @type{import('@popperjs/core').Placement[]} */
   const positions = ['top', 'left', 'right', 'bottom', 'bottom-start', 'bottom-end'];
 
   return (
